requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-jose==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
aiofiles==23.2.1
jinja2==3.1.3
python-jose[cryptography]==3.3.0
email-validator==2.1.0.post1
pytest==7.4.4
httpx==0.26.0  # For testing